Web.OIDC.Client

Description

OpenID Connect Discovery

module Web.OIDC.Client.Discovery

Settings and Tokens

data OIDC

This data type represents information needed in the OpenID flow.

newOIDC

Arguments

:: Provider
OP's information (obtained by discover)

-> OIDC

setCredentials

Arguments

:: ByteString
client ID

-> ByteString
client secret

-> ByteString
redirect URI (the HTTP endpont on your server that will receive a response from OP)

-> OIDC

-> OIDC

module Web.OIDC.Client.Tokens

Authorization Code Flow

module Web.OIDC.Client.CodeFlow

Types and Exceptions

module Web.OIDC.Client.Types

Re-exports