Web.OIDC.Client.CodeFlow

Description

Documentation

getAuthenticationRequestUrl

Arguments

:: (MonadThrow m, MonadCatch m)

=> OIDC

-> Scope
used to specify what are privileges requested for tokens. (use ScopeValue)

-> Maybe State
used for CSRF mitigation. (recommended parameter)

-> Parameters
Optional parameters

-> m URI

Make URL for Authorization Request.

requestTokens :: OIDC -> Code -> Manager -> IO Tokens

Request and validate tokens.

This function requests ID Token and Access Token to a OP's token endpoint, and validates the received ID Token. Returned Tokens value is a valid.

If a HTTP error has occurred or a tokens validation has failed, this function throws OpenIdException.

For testing

validateClaims :: Text -> Text -> IntDate -> JwtClaims -> IO ()

getCurrentIntDate :: IO IntDate