Web.OIDC.Client.Settings

Description

Documentation

data OIDC

This data type represents information needed in the OpenID flow.

Constructors

OIDC

Fields

oidcAuthorizationSeverUrl :: Text

oidcTokenEndpoint :: Text

oidcClientId :: ByteString

oidcClientSecret :: ByteString

oidcRedirectUri :: ByteString

oidcProvider :: Provider

def :: OIDC

newOIDC

Arguments

:: Provider
OP's information (obtained by discover)

-> OIDC

setCredentials

Arguments

:: ByteString
client ID

-> ByteString
client secret

-> ByteString
redirect URI (the HTTP endpont on your server that will receive a response from OP)

-> OIDC

-> OIDC