Web.OIDC.Client

Description

Client Obtains ID Token and Access Token

data OIDC

This data type represents information needed in the OpenID flow.

newOIDC :: CPRG g => IORef g -> OIDC

Create OIDC.

First argument is used in a token decoding on ID Token Validation.

newOIDC' :: OIDC

setProvider

Arguments

:: Provider
OP's information (obtain by discover)

-> OIDC

-> OIDC

setCredentials

Arguments

:: ByteString
client ID

-> ByteString
client secret

-> ByteString
redirect URI

-> OIDC

-> OIDC

getAuthenticationRequestUrl :: (MonadThrow m, MonadCatch m) => OIDC -> Scope -> Maybe State -> Parameters -> m URI

requestTokens :: OIDC -> Code -> Manager -> IO Tokens

Request and obtain valid tokens.

This function requests ID Token and Access Token to a OP's token endpoint, and validates the received ID Token. Returned value is a valid tokens.

Types

data Provider

An OpenID Provider information

type Scope = [ScopeValue]

data ScopeValue

Constructors

OpenId

Profile

Email

Address

Phone

OfflineAccess

type Code = ByteString

type State = ByteString

type Parameters = [(ByteString, Maybe ByteString)]

data Tokens

Constructors

Tokens

Fields

accessToken :: String

tokenType :: String

idToken :: IdToken

expiresIn :: Maybe Integer

refreshToken :: Maybe String

data IdToken

Constructors

IdToken

Fields

claims :: IdTokenClaims

jwt :: Jwt

data IdTokenClaims

Constructors

IdTokenClaims

Fields

iss :: String

sub :: String

aud :: [String]

exp :: IntDate

iat :: IntDate

Exception

data OpenIdException

Constructors

DiscoveryException String

InternalHttpException HttpException

JwtExceptoin JwtError

ValidationException String

Re-exports